+++ src/skrift-electron/renderer/components/Streams.tsx
@@ -2,13 +2,11 @@ import React from "react";
 import { StreamContainer } from "../containers/StreamContainer.js";
 import { Stream } from "../interfaces/state/index.js";
 import { DragDropStreamContainer } from "../containers/DragDropStreamContainer.js";
-import { OpenSearch } from "./OpenSearch.js";
 type Props = {
   streams: Stream[];
-  onOpenSearch: () => void;
 };
 
-export const Streams: React.FC<Props> = ({ streams, onOpenSearch }) => {
+export const Streams: React.FC<Props> = ({ streams }) => {
   return (
     <div className="h-screen flex flex-row justify-start text-sm bg-gray-200">
       {streams.map((stream, index) => (
@@ -16,7 +14,6 @@ export const Streams: React.FC<Props> = ({ streams, onOpenSearch }) => {
           <StreamContainer key={stream.key} path={[index]} stream={stream} />
         </DragDropStreamContainer>
       ))}
-      <OpenSearch onOpenSearch={onOpenSearch} />
     </div>
   );
 };