+++ src/skrift-electron/renderer/containers/StreamContainer.tsx
@@ -1,7 +1,8 @@
-import React from "react";
+import React, { useCallback, useContext } from "react";
 import { Stream } from "../components/Stream.js";
-import { Stream as StreamType } from "../interfaces/state/index.js";
+import { OpenCardMode, Stream as StreamType } from "../interfaces/state/index.js";
 import { StreamPath } from "../interfaces/path/index.js";
+import { StateContext } from "../interfaces/state/index.js";
 
 interface Props {
   path: StreamPath;
@@ -9,10 +10,20 @@ interface Props {
 }
 
 export const StreamContainer: React.FC<Props> = ({ path, stream }) => {
+  const [, { openCard }] = useContext(StateContext);
+
+  const handleOpenSearch = useCallback((mode: OpenCardMode) => {
+    openCard([...path, stream.cards.length], mode, {
+      type: "search",
+      query: "",
+    });
+  }, [openCard, path, stream.cards.length]);
+
   return (
     <Stream
       path={path}
       stream={stream}
+      onOpenSearch={handleOpenSearch}
     />
   );
 };