+++ src/skrift-electron/renderer/containers/StreamsContainer.tsx
@@ -1,16 +1,9 @@
-import React, { useCallback, useContext } from "react";
+import React, { useContext } from "react";
 import { StateContext } from "../interfaces/state/index.js";
 import { Streams } from "../components/Streams.js";
 
 export const StreamsContainer: React.FC = () => {
-  const [{ streams }, { openCard }] = useContext(StateContext);
+  const [{ streams }] = useContext(StateContext);
 
-  const handleOpenSearch = useCallback(() => {
-    openCard([streams.length - 1], "push", {
-      type: "search",
-      query: "",
-    });
-  }, [openCard, streams]);
-
-  return <Streams streams={streams} onOpenSearch={handleOpenSearch} />;
+  return <Streams streams={streams} />;
 };