+++ src/skrift-electron/renderer/interfaces/state/index.ts
@@ -106,7 +106,6 @@ export const State = {
       // Otherwise, path is a stream path, in which case the card should
       // be opened at the top.
       stream.cards.unshift(card);
-      return;
     }
 
     if (mode === "push") {
@@ -123,16 +122,14 @@ export const State = {
       const nextStream = state.streams[Path.stream(nextPath)];
 
       nextStream.cards.unshift(card);
-
-      return;
     }
 
     if (mode === "replace") {
       // Replace the card at the path with the new card.
       stream.cards.splice(Path.last(path), 1, card);
-
-      return;
     }
+
+    State.normalize(state);
   },
 
   updateCard<T extends Card>(state: State, path: Path, props: Partial<T>) {
@@ -261,13 +258,26 @@ export const State = {
   normalizeOnce(state: State): boolean {
     const { streams } = state;
 
-    // Use a for loop to allow early exit when an empty stream is removed.
-    for (let index = 0; index < streams.length; index++) {
-      if (streams[index].cards.length === 0) {
+    // There should always be at least 1 stream.
+    if (streams.length === 0) {
+      streams.push({
+        key: key++,
+        type: "stream",
+        cards: [],
+      });
+      return true;
+    }
+
+    // We always want at least 1 stream, empty or not.
+    if (streams.length === 1) return false;
+
+    // Find the first empty stream and remove it.
+    streams.forEach((stream, index) => {
+      if (stream.cards.length === 0) {
         streams.splice(index, 1);
         return true;
       }
-    }
+    });
 
     return false;
   },