+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Flip link click-logic: Click opens note below current note, Cmd/Ctrl-Click
   opens note in new stream to the right.
+- Move Search-launcher to below stream.
 
 ## 0.29.0 (2025-03-04)